+++ src/extension.js
@@ -207,6 +207,9 @@ function activate( context )
         var counts = provider.getTagCountsForStatusBar( fileFilter );
         var total = Object.values( counts ).reduce( function( a, b ) { return a + b; }, 0 );
 
+        var badgeTotal = config.shouldShowActivityBarBadge() ? total : 0;
+        todoTreeView.badge = { value: badgeTotal };
+
         var countRegex = new RegExp( "([^(]*)(\\(\\d+\\))*" );
         var match = countRegex.exec( todoTreeView.title );
         if( match !== null )
@@ -1680,6 +1683,10 @@ function activate( context )
                 {
                     rebuild();
                     documentChanged();
+                }
+                else if( e.affectsConfiguration( "todo-tree.general.showActivityBarBadge" ) )
+                {
+                    updateStatusBarAndTitleBar();
                 }
                 else
                 {